label-studio-backend/yolov5/requirements.txt
gunicorn==20.1.0
rq==1.10.1
torch>=1.13.1
torchvision>=0.8.1
tqdm>=4.64.0
pandas>=1.5.2
label-studio-ml>=1.0.8
opencv-python-headless>=4.6.0
matplotlib>=3.2.2
numpy>=1.18.5
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
gitpython
ipython
tensorboard>=2.4.1
seaborn>=0.11.0
